Move utility functions for testing to `test/mocha/helpers.js`.
