+++ lib/profileAgents.js
@@ -21,13 +21,8 @@ const {
   assertKeystoreOptions,
   createCapabilityAgent,
   getCollection,
-  getEdvConfig,
-  getEdvDocument,
   getPublicAliasTemplate,
-  parseEdvId,
-  removeSecretsFromRecords,
-  retryOperation,
-  timestampToDateString
+  removeSecretsFromRecords
 } = utils;
 const {config, util: {BedrockError}} = bedrock;
 
@@ -781,48 +776,6 @@ async function _refreshZcaps({record}) {
       record = await get({id: record.profileAgent.id, includeSecrets: true});
     }
   }
-
-  // FIXME: old -- needs to be reworked
-  /*
-  let zcapRefreshed = false;
-  for(const zcapName in zcaps) {
-    // cannot refresh this zcap
-    if(zcapName === 'profileCapabilityInvocationKey') {
-      continue;
-    }
-    // refresh zcap if it expires before/at refresh time
-    const zcap = zcaps[zcapName];
-    if(Date.parse(zcap.expires) <= refreshTime) {
-      if(zcapName === 'userDocument') {
-        // other zcaps to refresh can be found in the EDV user doc
-        await _refreshProfileAgentUserDocZcaps({zcap, now, record});
-      }
-      zcap.expires = timestampToDateString(now + EDV_ZCAP_TTL);
-      zcapRefreshed = true;
-    }
-  }
-
-  if(zcapRefreshed) {
-    // update the profile agent record
-    record.profileAgent.sequence += 1;
-    const res = await retryOperation(async () => {
-      return update({profileAgent: record.profileAgent});
-    }, async () => {
-      return get({
-        id: record.profileAgent.id, includeSecrets: true
-      });
-    });
-    const recordUpdated = res?.lastErrorObject?.updatedExisting;
-    if(recordUpdated) {
-      record = res.value;
-    } else {
-      // delete _id
-      delete res._id;
-      record = res;
-    }
-  }*/
-
-  return record;
 }
 
 async function _generateProfileAgentZcaps({profileSigner, now, zcaps}) {
@@ -841,71 +794,6 @@ async function _generateProfileAgentZcaps({profileSigner, now, zcaps}) {
   }
   return {zcaps: result};
 }
-
-// FIXME: remove
-/*
-async function _refreshProfileAgentUserDocZcaps({zcap, now, record} = {}) {
-  let zcapRefreshed = false;
-  const {zcapRefreshThreshold} = config.profile.profileAgent;
-  const kmsClient = new KmsClient({httpsAgent});
-  const profileSigner = await getProfileSigner({
-    kmsClient, profileAgentRecord: record
-  });
-  const edvId = parseEdvId({capability: zcap});
-  const edvClient = new EdvClient({id: edvId, httpsAgent, keyResolver});
-  const edvConfig = await getEdvConfig({edvClient, profileSigner});
-  const docId = zcap.invocationTarget.split('/').pop();
-  // get profile agent user doc
-  const profileAgentUserDoc = await getEdvDocument({
-    docId, edvConfig, edvClient, kmsClient, profileSigner
-  });
-  console.log('profile agent user doc',
-    JSON.stringify(profileAgentUserDoc.content, null, 2));
-
-  // FIXME: new, replace above
-  {
-    const userDoc = await _readUserEdvDocument({profileAgentRecord: record});
-    console.log('*****NEW profile agent user doc',
-      JSON.stringify(userDoc.content, null, 2));
-  }
-
-  const updatedProfileAgentUserDoc = klona(profileAgentUserDoc);
-  const {zcaps: profileAgentUserDocZcaps} =
-    updatedProfileAgentUserDoc.content;
-  for(const profileAgentUserDocZcapName in profileAgentUserDocZcaps) {
-    if(profileAgentUserDocZcapName === 'profileCapabilityInvocationKey') {
-      continue;
-    }
-    const profileAgentUserDocZcap =
-      profileAgentUserDocZcaps[profileAgentUserDocZcapName];
-    const {
-      expires: profileAgentUserDocZcapExpires
-    } = profileAgentUserDocZcap;
-    const profileAgentUserDocZcapExpTime =
-      Date.parse(profileAgentUserDocZcapExpires);
-    // check if the zcap is close to expiry
-    if((profileAgentUserDocZcapExpTime - now) <= zcapRefreshThreshold) {
-      // refresh the zcap's expires property
-      profileAgentUserDocZcap.expires = timestampToDateString(
-        now + EDV_ZCAP_TTL);
-      zcapRefreshed = true;
-    }
-  }
-  if(zcapRefreshed) {
-    return retryOperation(async () => {
-      return edvClient.update({
-        doc: updatedProfileAgentUserDoc,
-        invocationSigner: profileSigner,
-        keyResolver
-      });
-    }, async () => {
-      return getEdvDocument({
-        docId, edvConfig, edvClient, kmsClient, profileSigner
-      });
-    });
-  }
-}
-*/
 
 async function _getUserEdvDocument({profileAgentRecord, zcaps, kmsClient}) {
   const invocationSigner = await getSigner({profileAgentRecord});
