+++ lib/utils.js
@@ -3,7 +3,6 @@
  */
 import * as bedrock from '@bedrock/core';
 import * as database from '@bedrock/mongodb';
-import {Hmac, KeyAgreementKey} from '@digitalbazaar/webkms-client';
 import assert from 'assert-plus';
 import {CapabilityAgent} from '@digitalbazaar/webkms-client';
 import {CryptoLD} from 'crypto-ld';
@@ -14,7 +13,6 @@ import {
 import {
   Ed25519VerificationKey2020
 } from '@digitalbazaar/ed25519-verification-key-2020';
-import {EdvDocument} from '@digitalbazaar/edv-client';
 import {getAppIdentity} from '@bedrock/app-identity';
 import {httpsAgent} from '@bedrock/https-agent';
 import {promisify} from 'node:util';
@@ -107,75 +105,12 @@ export function parseKeystoreId(keyId) {
   return keyId.slice(0, idx);
 }
 
-export async function getEdvConfig({edvClient, profileSigner} = {}) {
-  return edvClient.getConfig({
-    invocationSigner: profileSigner
-  });
-}
-
-// FIXME: remove if possible
-export async function getEdvDocument({
-  docId, edvConfig, edvClient, kmsClient, profileSigner
-} = {}) {
-  const {hmac, keyAgreementKey} = edvConfig;
-
-  const doc = new EdvDocument({
-    invocationSigner: profileSigner,
-    id: docId,
-    keyAgreementKey: new KeyAgreementKey({
-      id: keyAgreementKey.id,
-      type: keyAgreementKey.type,
-      invocationSigner: profileSigner,
-      kmsClient
-    }),
-    hmac: new Hmac({
-      id: hmac.id,
-      type: hmac.type,
-      invocationSigner: profileSigner,
-      kmsClient
-    }),
-    client: edvClient
-  });
-  return doc.read();
-}
-
 export function removeSecretsFromRecords({records}) {
   return records.map(record => {
     // eslint-disable-next-line no-unused-vars
     const {secrets, ...rest} = record;
     return rest;
   });
-}
-
-export function parseEdvId({capability}) {
-  const {invocationTarget} = capability;
-  const idx = invocationTarget.lastIndexOf('/documents');
-  if(idx === -1) {
-    throw new Error(
-      `Invalid EDV invocation target (${invocationTarget}).`);
-  }
-  return invocationTarget.slice(0, idx);
-}
-
-export async function retryOperation(operationFunction, onInvalidStateError) {
-  const retry = true;
-  while(retry) {
-    try {
-      const result = await operationFunction();
-      return result;
-    } catch(e) {
-      const status = e?.status || e?.cause?.status;
-      if(status === 409 || e.name == 'InvalidStateError') {
-        const result = onInvalidStateError();
-        return result;
-      }
-    }
-  }
-}
-
-// timestamp is in milliseconds
-export function timestampToDateString(timestamp) {
-  return new Date(timestamp).toISOString().slice(0, -5) + 'Z';
 }
 
 /**
